CSS styling in some very inefficient workarounds. Flexbox be acting up sometimes ngl
