+++ src/App.tsx
@@ -1,14 +1,14 @@
 import './App.css'
-import { Title } from "./title.tsx";
-import { Intro } from "./intro.tsx";
-import { Hobby1} from "./hobby1.tsx";
-import { List1 } from "./list1.tsx"
-import { Table1 } from "./table1.tsx";
-import { Form1 } from "./form1.tsx";
-import { Hobby2 } from "./hobby2.tsx";
-import { List2 } from "./list2.tsx";
-import { Table2 } from "./table2.tsx";
-import { Form2 } from "./form2.tsx";
+import { Title } from "./Components/title.tsx";
+import { Intro } from "./Components/intro.tsx";
+import { Hobby1} from "./Components/hobby1.tsx";
+import { List1 } from "./Components/list1.tsx"
+import { Table1 } from "./Components/table1.tsx";
+import { Form1 } from "./Components/form1.tsx";
+import { Hobby2 } from "./Components/hobby2.tsx";
+import { List2 } from "./Components/list2.tsx";
+import { Table2 } from "./Components/table2.tsx";
+import { Form2 } from "./Components/form2.tsx";
 
 function App() {
 
@@ -16,17 +16,26 @@ function App() {
     <>
         <Title />
         <Intro name="Patrick Tirch" name2="William Hanlon" />
+        <br/>
         <div className="bothHobbies">
             <div className="PatrickHobby">
                 <Hobby1 />
+                <hr/>
                 <List1 />
+                <hr/>
                 <Table1 />
+                <hr/>
                 <Form1 />
             </div>
+            <br/>
+            <hr/>
             <div className="WilliamHobby">
                 <Hobby2 />
+                <hr/>
                 <List2 />
+                <hr/>
                 <Table2 />
+                <hr/>
                 <Form2 />
             </div>
         </div>
