+++ src/Components/Image.css
@@ -1,0 +1,14 @@
+img {
+    max-width: 100%;
+    border-radius: 10px;
+    justify-content: center;
+    align-items: center;
+}
+
+.center-div{
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin: 1rem;
+
+}
